[REFACTOR] 패키지 이전: UserAvatar -> avatar/model
