+++ src/main/java/com/wakeUpTogetUp/togetUp/api/avatar/model/UserAvatar.java
@@ -1,7 +1,6 @@
-package com.wakeUpTogetUp.togetUp.api.users.model;
+package com.wakeUpTogetUp.togetUp.api.avatar.model;
 
-import com.wakeUpTogetUp.togetUp.api.avatar.model.Avatar;
-
+import com.wakeUpTogetUp.togetUp.api.users.model.User;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -12,7 +11,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
